images portion done forgot password and ui fixing of both dashboards done

+++ controller/servicesController.js
@@ -1,9 +1,86 @@
 const Service = require("../models/serviceModel");
+const multer = require("multer");
+const path = require("path");
+const cloudinary = require("../utils/cloudinaryConfig");
+const { CloudinaryStorage } = require("multer-storage-cloudinary");
+
+const storage = new CloudinaryStorage({
+  cloudinary: cloudinary,
+  params: {
+    folder: "services",
+    format: async (req, file) => {
+      // Get the file extension
+      const fileExtension = path.extname(file.originalname).toLowerCase();
+
+      // Check if the extension is allowed and return the format
+      if (fileExtension === ".jpg" || fileExtension === ".jpeg") {
+        return "jpg";
+      } else if (fileExtension === ".png") {
+        return "png";
+      } else {
+        throw new Error("Unsupported file format");
+      }
+    },
+    public_id: (req, file) => {
+      // Remove file extension and add a unique identifier to the public ID
+      const uniqueID = Date.now();
+      return `services/${path.parse(file.originalname).name}_${uniqueID}`;
+    },
+  },
+});
+
+const parser = multer({ storage: storage });
 
 const getAllServices = async (req, res, next) => {
   try {
     const services = await Service.find({});
-    res.json(services);
+
+    // Fetch the thumbnail URLs from Cloudinary
+    const servicesWithThumbnails = await Promise.all(
+      services.map(async (service) => {
+        try {
+          const thumbnailData = await cloudinary.api.resource(
+            service.thumbnail,
+            {
+              resource_type: "image",
+            }
+          );
+
+          const fetchedImages = await Promise.all(
+            service.images.map(async (publicId) => {
+              try {
+                return await cloudinary.api.resource(publicId, {
+                  resource_type: "image",
+                });
+              } catch (error) {
+                console.error(
+                  `Error fetching image with public ID ${publicId}:`,
+                  error
+                );
+                return null;
+              }
+            })
+          );
+
+          const validImagesData = fetchedImages.filter(imageUrl => imageUrl !== null);
+
+          // Return the service data with the fetched thumbnail URL
+          return {
+            ...service._doc,
+            thumbnail: thumbnailData.secure_url,
+            images: validImagesData.map(imageData => imageData.secure_url)
+          };
+        } catch (error) {
+          console.error(
+            `Error fetching thumbnail with public ID ${service.thumbnail}:`,
+            error
+          );
+          return service;
+        }
+      })
+    );
+
+    res.json(servicesWithThumbnails);
   } catch (error) {
     res.status(500).json({ message: `service stuck here` });
     next(error);
@@ -12,37 +89,105 @@ const getAllServices = async (req, res, next) => {
 
 const getAllServiceIds = async (req, res, next) => {
   try {
-    const servicesIds = await Service.find({}, {_id:1});
+    const servicesIds = await Service.find({}, { _id: 1 });
     res.json(servicesIds);
   } catch (error) {
     next(error);
   }
 };
 
-const getServiceById = async (req, res, next) =>{
- try{
-  const serviceid = req.url.toString().split('/');
-  const serviceById = await Service.findById(serviceid[2]);
-  res.json(serviceById);
- }catch(error){
-  next(error)
- }
-}
+const getServiceById = async (req, res, next) => {
+  try {
+    const serviceid = req.url.toString().split("/");
+    const serviceById = await Service.findById(serviceid[2]);
+    res.json(serviceById);
+  } catch (error) {
+    next(error);
+  }
+};
 
-const addNewService = async (req, res, next) => {
+const getServiceDataAndImages = async (req, res, next) => {
+  try {
+    const serviceId = req.params.id;
+    const service = await Service.findById(serviceId);
+
+    if (!service) {
+      res.status(404).json({ message: "Service not found" });
+      return;
+    }
+
+    const fetchedImages = await Promise.all(
+      service.images.map(async (publicId) => {
+        try {
+          return await cloudinary.api.resource(publicId, {
+            resource_type: "image",
+          });
+        } catch (error) {
+          console.error(
+            `Error fetching image with public ID ${publicId}:`,
+            error
+          );
+          return null;
+        }
+      })
+    );
+
+    console.log("in here 2");
+    const serviceDataWithImages = {
+      ...service._doc,
+      images: fetchedImages
+        .filter((img) => img !== null)
+        .map((img) => img.secure_url),
+    };
+    console.log("in here 3");
+    console.log(serviceDataWithImages);
+
+    res.json(serviceDataWithImages);
+  } catch (error) {
+    next(error);
+  }
+};
+
+
+const addNewServiceWithImages = async (req, res, next) => {
   const { title, description, newPackages } = req.body;
   const packages = JSON.parse(newPackages);
+
+  const thumbnailImage =
+    req.files.thumbnail.length > 0 ? req.files.thumbnail[0] : undefined;
+  const images = req.files.images.map((file) => file.filename);
+
   const parsedService = {
     title,
     description,
     packages,
+    images,
+    thumbnail: thumbnailImage.filename,
+  };
+
+  const newService = new Service(parsedService);
+  try {
+    const savedService = await newService.save();
+    res.status(200).json({ message: `Service is Successfully Saved!!` });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const addNewService = async (req, res, next) => {
+  const { title, description, newPackages, images } = req.body;
+  const packages = JSON.parse(newPackages);
+  const imagesArray = JSON.parse(images);
+  const parsedService = {
+    title,
+    description,
+    packages,
+    images: imagesArray,
   };
   const newService = new Service(parsedService);
   try {
     const savedService = await newService.save();
-    res
-      .status(200)
-      .json({ message: `Service is Successfully Saved!!` });
+    res.status(200).json({ message: `Service is Successfully Saved!!` });
   } catch (error) {
     next(error);
   }
@@ -50,13 +195,33 @@ const addNewService = async (req, res, next) => {
 
 const deleteService = async (req, res, next) => {
   try {
-    const serviceid = req.url.toString().split('/');
+    const serviceid = req.url.toString().split("/");
+    const serviceToDelete = await Service.findById(serviceid[2]);
+
+    if (!serviceToDelete) {
+      res.status(404).json({ message: "Service not found" });
+      return;
+    }
+
+    // Delete the service's images from Cloudinary
+    const deletePromises = serviceToDelete.images.map((publicId) => {
+      return cloudinary.uploader.destroy(publicId);
+    });
+
+    // Delete the service's thumbnail from Cloudinary
+    if (serviceToDelete.thumbnail) {
+      deletePromises.push(cloudinary.uploader.destroy(serviceToDelete.thumbnail));
+    }
+
+    await Promise.all(deletePromises);
+
     const deletedService = await Service.findByIdAndDelete(serviceid[2]);
+
     res.json(deletedService);
   } catch (error) {
     next(error);
   }
-}
+};
 
 const updateServiceById = async (req, res, next) => {
   const { title, description, newPackages } = req.body;
@@ -65,22 +230,41 @@ const updateServiceById = async (req, res, next) => {
     title,
     description,
     packages,
-  }
+  };
   try {
-    const serviceid = req.url.toString().split('/');
-    const updatedService = await Service.findByIdAndUpdate(serviceid[2], newPackage);
+    const serviceid = req.url.toString().split("/");
+    const updatedService = await Service.findByIdAndUpdate(
+      serviceid[2],
+      newPackage
+    );
     res.json(updatedService);
   } catch (error) {
     next(error);
   }
-}
+};
 
+const uploadServiceImage = async (req, res, next) => {
+  try {
+    const result = await cloudinary.uploader.upload(req.file.path);
+    res.status(200).json({
+      imageUrl: result.secure_url,
+      public_id: result.public_id,
+    });
+  } catch (error) {
+    res.status(500).json({ message: `Error uploading image` });
+    next(error);
+  }
+};
 
 module.exports = {
   getAllServices,
   getAllServiceIds,
   getServiceById,
+  getServiceDataAndImages,
   addNewService,
+  addNewServiceWithImages,
   deleteService,
-  updateServiceById
+  updateServiceById,
+  uploadServiceImage,
+  parser,
 };
